--- a/Configuration/Eras/python/Modifier_run2_HCAL_2017_cff.py
+++ b/Configuration/Eras/python/Modifier_run2_HCAL_2017_cff.py
@@ -1,0 +1,4 @@
+import FWCore.ParameterSet.Config as cms
+
+run2_HCAL_2017 =  cms.Modifier()
+