--- a/Configuration/Eras/python/Modifier_run2_HCAL_2017_cff.py
+++ b/Configuration/Eras/python/Modifier_run2_HCAL_2017_cff.py
@@ -1,4 +1,8 @@
 import FWCore.ParameterSet.Config as cms
 
+# This modifier should be used for "common" HCAL 2017 customizations
+# Initially, this means the packer and DQM
+# subdetector-specific changes for sim, reco, etc. should go in run2_HE_2017 or run2_HF_2017
+
 run2_HCAL_2017 =  cms.Modifier()