--- a/Configuration/Eras/python/Modifier_run2_HF_2017_cff.py
+++ b/Configuration/Eras/python/Modifier_run2_HF_2017_cff.py
@@ -1,4 +1,6 @@
 import FWCore.ParameterSet.Config as cms
 
+# This modifier is for HF-specific changes for sim, reco, etc.
+
 run2_HF_2017 =  cms.Modifier()