--- a/Configuration/Geometry/python/GeometryExtended2017newReco_cff.py
+++ b/Configuration/Geometry/python/GeometryExtended2017newReco_cff.py
@@ -1,0 +1,33 @@
+import FWCore.ParameterSet.Config as cms
+
+# Ideal geometry, needed for transient ECAL alignement
+from Configuration.Geometry.GeometryExtended2017new_cff import *
+
+
+# Reconstruction geometry services
+#  Tracking Geometry
+#bah - well, this is not a cfi!
+from Geometry.CommonDetUnit.globalTrackingGeometry_cfi import *
+
+#Tracker
+from RecoTracker.GeometryESProducer.TrackerRecoGeometryESProducer_cfi import *
+from Geometry.TrackerNumberingBuilder.trackerTopology_cfi import *
+
+#Muon
+from Geometry.MuonNumbering.muonNumberingInitialization_cfi import *
+from RecoMuon.DetLayers.muonDetLayerGeometry_cfi import *
+
+#  Alignment
+from Geometry.TrackerGeometryBuilder.idealForDigiTrackerGeometry_cff import *
+from Geometry.CSCGeometryBuilder.idealForDigiCscGeometry_cff import *
+from Geometry.DTGeometryBuilder.idealForDigiDtGeometry_cff import *
+trackerGeometry.applyAlignment = False
+
+#  Calorimeters
+from Geometry.CaloEventSetup.CaloTopology_cfi import *
+from Geometry.CaloEventSetup.CaloGeometry_cff import *
+from Geometry.CaloEventSetup.EcalTrigTowerConstituents_cfi import *
+from Geometry.EcalMapping.EcalMapping_cfi import *
+from Geometry.EcalMapping.EcalMappingRecord_cfi import *
+from Geometry.HcalCommonData.hcalDDDRecConstants_cfi import *
+from Geometry.HcalEventSetup.hcalTopologyIdeal_cfi import *