--- a/Configuration/Geometry/python/GeometryExtended2017new_cff.py
+++ b/Configuration/Geometry/python/GeometryExtended2017new_cff.py
@@ -1,0 +1,11 @@
+import FWCore.ParameterSet.Config as cms
+
+#
+# Geometry master configuration
+#
+# Ideal geometry, needed for simulation
+from Geometry.CMSCommonData.cmsExtendedGeometry2017newXML_cfi import *
+from Geometry.TrackerNumberingBuilder.trackerNumberingGeometry_cfi import *
+from Geometry.HcalCommonData.hcalParameters_cfi      import *
+from Geometry.HcalCommonData.hcalDDDSimConstants_cfi import *
+