--- a/Configuration/PyReleaseValidation/python/relval_2017.py
+++ b/Configuration/PyReleaseValidation/python/relval_2017.py
@@ -15,8 +15,8 @@ workflows = Matrix()
 
 numWFStart=10000
 numWFSkip=200
-#2017 WFs to run in IB (TenMuE_0_200, TTbar, ZEE, MinBias, TTbar PU, ZEE PU,TTbar NewPIX,TTbar HCALdev)
-numWFIB = [10021.0,10024.0,10025.0,10026.0,10023.0,10224.0,10225.0,10424.0,10624.0]
+#2017 WFs to run in IB (TenMuE_0_200, TTbar, ZEE, MinBias, TTbar PU, ZEE PU,TTbar NewPIX,TTbar HCALdev,TTbar 2017new)
+numWFIB = [10021.0,10024.0,10025.0,10026.0,10023.0,10224.0,10225.0,10424.0,10624.0,10824.0]
 for i,key in enumerate(upgradeKeys[2017]):
     numWF=numWFStart+i*numWFSkip
     for frag in upgradeFragments: