--- a/Configuration/PyReleaseValidation/python/upgradeWorkflowComponents.py
+++ b/Configuration/PyReleaseValidation/python/upgradeWorkflowComponents.py
@@ -69,7 +69,7 @@ upgradeProperties[2017] = {
         'ScenToRun' : ['GenSimFull','DigiFull','RecoFull','HARVESTFull'],
     },
     '2017HCALdev' : {
-        'Geom' : 'Extended2017dev',
+        'Geom' : 'DB:Extended',
         'GT' : 'auto:phase1_2017_hcaldev',
         'Era' : 'Run2_2017_HCALdev',
         'ScenToRun' : ['GenSimFull','DigiFull','RecoFull','HARVESTFull'],