--- a/Configuration/StandardSequences/python/Eras.py
+++ b/Configuration/StandardSequences/python/Eras.py
@@ -18,6 +18,7 @@ class Eras (object):
                  'Run2_2017',
                  'Run2_2017_NewFPix',
                  'Run2_2017_HCALdev',
+                 'Run2_2017_new',
                  'Run2_2017_trackingRun2',
                  'Run2_2017_trackingPhase1PU70',
                  'Run3',