--- a/Configuration/StandardSequences/python/GeometryConf.py
+++ b/Configuration/StandardSequences/python/GeometryConf.py
@@ -22,6 +22,7 @@ GeometryConf={
     'Extended2015dev': 'Extended2015dev,Extended2015devReco',
     'Extended2016'   : 'Extended2016,Extended2016Reco',
     'Extended2017'   : 'Extended2017,Extended2017Reco',
+    'Extended2017dev': 'Extended2017dev,Extended2017devReco',
     'Extended2019'   : 'Extended2019,Extended2019Reco',
     'ExtendedGFlash' : 'ExtendedGFlash,ExtendedGFlashReco',
     'All'         : 'Configuration.Geometry.GeometrySimAll_cff,Reco',