--- a/DQM/HcalTasks/python/OfflineSourceSequence_hi.py
+++ b/DQM/HcalTasks/python/OfflineSourceSequence_hi.py
@@ -7,12 +7,14 @@ import FWCore.ParameterSet.Config as cms
 
 #	import the tasks
 from DQM.HcalTasks.DigiTask import digiTask
+from DQM.HcalTasks.DigiPhase1Task import digiPhase1Task
 from DQM.HcalTasks.RawTask import rawTask
 from DQM.HcalTasks.TPTask import tpTask
 from DQM.HcalTasks.RecHitTask import recHitTask
 
 #	set processing type to Offine
 digiTask.ptype = cms.untracked.int32(1)
+from DQM.HcalTasks.DigiPhase1Task import digiPhase1Task
 tpTask.ptype = cms.untracked.int32(1)
 recHitTask.ptype = cms.untracked.int32(1)
 rawTask.ptype = cms.untracked.int32(1)
@@ -36,4 +38,10 @@ hcalOfflineSourceSequence = cms.Sequence(
 	+tpTask
 	+recHitTask
 	+rawTask)
+
+_phase1_hcalOfflineSourceSequence = hcalOfflineSourceSequence.copy() 
+_phase1_hcalOfflineSourceSequence.insert(_phase1_hcalOfflineSourceSequence.index(digiTask),digiPhase1Task)
+
+from Configuration.StandardSequences.Eras import eras
+eras.run2_HCAL_2017.toReplaceWith( hcalOfflineSourceSequence, _phase1_hcalOfflineSourceSequence )