--- a/DQM/HcalTasks/python/OfflineSourceSequence_pp.py
+++ b/DQM/HcalTasks/python/OfflineSourceSequence_pp.py
@@ -7,12 +7,14 @@ import FWCore.ParameterSet.Config as cms
 
 #	import the tasks
 from DQM.HcalTasks.DigiTask import digiTask
+from DQM.HcalTasks.DigiPhase1Task import digiPhase1Task
 from DQM.HcalTasks.RawTask import rawTask
 from DQM.HcalTasks.TPTask import tpTask
 from DQM.HcalTasks.RecHitTask import recHitTask
 
 #	set processing type to Offine
 digiTask.ptype = cms.untracked.int32(1)
+digiPhase1Task.ptype = cms.untracked.int32(1)
 tpTask.ptype = cms.untracked.int32(1)
 recHitTask.ptype = cms.untracked.int32(1)
 rawTask.ptype = cms.untracked.int32(1)
@@ -26,3 +28,8 @@ hcalOfflineSourceSequence = cms.Sequence(
 	+recHitTask
 	+rawTask)
 
+_phase1_hcalOfflineSourceSequence = hcalOfflineSourceSequence.copy()
+_phase1_hcalOfflineSourceSequence.insert(_phase1_hcalOfflineSourceSequence.index(digiTask),digiPhase1Task)
+
+from Configuration.StandardSequences.Eras import eras
+eras.run2_HCAL_2017.toReplaceWith( hcalOfflineSourceSequence, _phase1_hcalOfflineSourceSequence )