--- a/RecoLocalCalo/Configuration/python/hcalLocalReco_cff.py
+++ b/RecoLocalCalo/Configuration/python/hcalLocalReco_cff.py
@@ -20,3 +20,15 @@ eras.phase2_hcal.toReplaceWith( hfreco, _hfUpgradeReco )
 eras.phase2_hcal.toModify( hfreco, digiLabel = cms.InputTag('simHcalDigis','HFUpgradeDigiCollection') )
 eras.phase2_hcal.toModify( zdcreco, digiLabel = cms.InputTag('simHcalUnsuppressedDigis'), digiLabelhcal = cms.InputTag('simHcalUnsuppressedDigis') )
 eras.phase2_hcal.toReplaceWith( hcalLocalRecoSequence, _phase2_hcalLocalRecoSequence )
+
+from RecoLocalCalo.HcalRecProducers.hfprereco_cfi import hfprereco
+from RecoLocalCalo.HcalRecProducers.HFPhase1Reconstructor_cfi import hfreco as _phase1_hfreco
+from RecoLocalCalo.HcalRecProducers.HBHEPhase1Reconstructor_cfi import hbheprereco as _phase1_hbheprereco
+
+_phase1_hcalLocalRecoSequence = hcalLocalRecoSequence.copy()
+_phase1_hcalLocalRecoSequence.insert(_phase1_hcalLocalRecoSequence.index(hbheprereco),hfprereco)
+
+eras.run2_HF_2017.toReplaceWith( hcalLocalRecoSequence, _phase1_hcalLocalRecoSequence )
+eras.run2_HF_2017.toReplaceWith( hfreco, _phase1_hfreco )
+eras.run2_HE_2017.toReplaceWith( hbheprereco, _phase1_hbheprereco )
+eras.run2_HE_2017.toModify( hbheprereco, saveInfos = cms.bool(True) )