--- a/RecoLocalCalo/HcalRecProducers/python/hfprereco_cfi.py
+++ b/RecoLocalCalo/HcalRecProducers/python/hfprereco_cfi.py
@@ -1,7 +1,7 @@
 import FWCore.ParameterSet.Config as cms
 
 hfprereco = cms.EDProducer("HFPreReconstructor",
-    digiLabel = cms.InputTag("hcalDigis", "HFQIE10DigiCollection"),
+    digiLabel = cms.InputTag("hcalDigis"),
     dropZSmarkedPassed = cms.bool(True),
     tsFromDB = cms.bool(False)
 )