Long-anticipated... thanks to @mmusich for the autocond in #15884.

This PR adds two new workflows: 2017HCALdev ("old" pixel with new HCAL geometry/conditions/customizations for 2017) and 2017AllNew ("new" pixel with new HCAL).

Tested using workflows 10624.0, 10824.0 (ttbar, HCALdev and AllNew respectively).

@igv4321, @mariadalfonso, let me know if I missed any reco settings you want to have turned on by default.

attn: @abdoulline, @lihux25, @hatakeyamak, @pdudero