@cmsbuild 2016-09-17
A new Pull Request was created by  @kpedro88 (Kevin Pedro) for CMSSW_8_1_X.

It involves the following packages:

Configuration/Eras
Configuration/PyReleaseValidation
Configuration/StandardSequences
DQM/HcalTasks
EventFilter/HcalRawToDigi
EventFilter/RawDataCollector
RecoLocalCalo/Configuration
RecoLocalCalo/HcalRecProducers


@dmitrijus, @cvuosalo, @emeschi, @franzoni, @fabozzi, @cmsbuild, @srimanob, @slava77, @mommsen, @vanbesien, @hengne, @davidlange6 can you please review it and eventually sign? Thanks.
@ghellwig, @makortel, @GiacomoSguazzoni, @rovere, @VinInn, @Martin-Grunewald, @deguio, @mariadalfonso, @dgulhan this is something you requested to watch as well.
@slava77, @smuzaffar you are the release manager for this.

cms-bot commands are list here cms-sw/cmssw#13028