@cmsbuild 2016-09-17
-1

Tested at: a19eefbaf71e4076b31464765b01cee20517b8c4


The following merge commits were also included on top of IB + this PR after doing git cms-merge-topic: 
https://github.com/cms-sw/cmssw/commit/813e5adfbc1678f6e7f1f7558b9cb2bc8654a8b8
https://github.com/cms-sw/cmssw/commit/51dd54a33588b0fa3e3620cc32c91c0c7c62ca93
https://github.com/cms-sw/cmssw/commit/12077f84fef9dc371adcedba4c3ec694bf563c4d
https://github.com/cms-sw/cmssw/commit/851461bd79ace5087f390d4f3c5ffb2cc92b555d
https://github.com/cms-sw/cmssw/commit/5cb0fc27e76172acf3aeb512cad31ad3fb0ba42b
https://github.com/cms-sw/cmssw/commit/43dfd1b14bed7dae04e351326c0846ae44be3bdb
https://github.com/cms-sw/cmssw/commit/032ad55306bc243fab10501289571eb9adf00679
https://github.com/cms-sw/cmssw/commit/3cfd184c0828e5546fb67c5d8c0274a0618430d1
https://github.com/cms-sw/cmssw/commit/4c378a3a0065db1ef2c2a6cf574ff1468c8803cb
https://github.com/cms-sw/cmssw/commit/f5690193ddb3d9d8fe40d664a446407df12077af
https://github.com/cms-sw/cmssw/commit/941fa6eeff293377ca537cc7449c928e4dccc7bc
https://github.com/cms-sw/cmssw/commit/a53847c15bfc1e8bf99fa96c286eadcd22b1ff4e
https://github.com/cms-sw/cmssw/commit/a03ca88e7099d97a5bd8455c3f2fb1e2fcd09b28
https://github.com/cms-sw/cmssw/commit/bc0c67977f52dcf7888761af8c4d0b03bf9e5b34
https://github.com/cms-sw/cmssw/commit/af763d9cffd2759e1fad5a46ba4c8b0a59854188
https://github.com/cms-sw/cmssw/commit/ec5ff3cd8da9019b359cc01f39002ae9637f6f15
You can see more details here:
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-15889/15233/git-log-recent-commits
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-15889/15233/git-merge-result

You can see the results of the tests here:
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-15889/15233/summary.html

I found follow errors while testing this PR

**Failed tests:  RelVals**

* **RelVals**:

When I ran the RelVals I found an error in the following worklfows:
20824.0 step3
<pre>runTheMatrix-results/20824.0_TTbar_13+TTbar_13TeV_TuneCUETP8M1_2023D3_GenSimFull+DigiFull_2023D3+RecoFullGlobal_2023D3+HARVESTFullGlobal_2023D3/step3_TTbar_13+TTbar_13TeV_TuneCUETP8M1_2023D3_GenSimFull+DigiFull_2023D3+RecoFullGlobal_2023D3+HARVESTFullGlobal_2023D3.log

</pre>


The following merge commits were also included on top of IB + this PR after doing git cms-merge-topic: 
https://github.com/cms-sw/cmssw/commit/813e5adfbc1678f6e7f1f7558b9cb2bc8654a8b8
https://github.com/cms-sw/cmssw/commit/51dd54a33588b0fa3e3620cc32c91c0c7c62ca93
https://github.com/cms-sw/cmssw/commit/12077f84fef9dc371adcedba4c3ec694bf563c4d
https://github.com/cms-sw/cmssw/commit/851461bd79ace5087f390d4f3c5ffb2cc92b555d
https://github.com/cms-sw/cmssw/commit/5cb0fc27e76172acf3aeb512cad31ad3fb0ba42b
https://github.com/cms-sw/cmssw/commit/43dfd1b14bed7dae04e351326c0846ae44be3bdb
https://github.com/cms-sw/cmssw/commit/032ad55306bc243fab10501289571eb9adf00679
https://github.com/cms-sw/cmssw/commit/3cfd184c0828e5546fb67c5d8c0274a0618430d1
https://github.com/cms-sw/cmssw/commit/4c378a3a0065db1ef2c2a6cf574ff1468c8803cb
https://github.com/cms-sw/cmssw/commit/f5690193ddb3d9d8fe40d664a446407df12077af
https://github.com/cms-sw/cmssw/commit/941fa6eeff293377ca537cc7449c928e4dccc7bc
https://github.com/cms-sw/cmssw/commit/a53847c15bfc1e8bf99fa96c286eadcd22b1ff4e
https://github.com/cms-sw/cmssw/commit/a03ca88e7099d97a5bd8455c3f2fb1e2fcd09b28
https://github.com/cms-sw/cmssw/commit/bc0c67977f52dcf7888761af8c4d0b03bf9e5b34
https://github.com/cms-sw/cmssw/commit/af763d9cffd2759e1fad5a46ba4c8b0a59854188
https://github.com/cms-sw/cmssw/commit/ec5ff3cd8da9019b359cc01f39002ae9637f6f15
You can see more details here:
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-15889/15233/git-log-recent-commits
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-15889/15233/git-merge-result