@cmsbuild 2016-09-17
The tests are being triggered in jenkins.