@cmsbuild 2016-09-18
-1

Tested at: 15e780dfa01db9c7b9f9de7aeb428c9580987348

You can see the results of the tests here:
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-15889/15246/summary.html

I found follow errors while testing this PR

**Failed tests:  Build**

* **Build**:

I found an error when building: 
 
 <pre>/build/cmsbuild/jenkins-workarea/workspace/ib-any-integration/CMSSW_8_1_X_2016-09-17-1100/src/EventFilter/CSCRawToDigi/interface/CSCCLCTData.h:3:61: warning: /cvmfs/cms-ib.cern.ch/week0/slc6_amd64_gcc530/cms/cmssw-patch/CMSSW_8_1_X_2016-09-17-1100/src/DataFormats/CSCDigi/interface/CSCComparatorDigi.h is shorter than expected
In file included from /build/cmsbuild/jenkins-workarea/workspace/ib-any-integration/CMSSW_8_1_X_2016-09-17-1100/src/EventFilter/CSCRawToDigi/interface/CSCTMBData.h:12:0,
                 from /build/cmsbuild/jenkins-workarea/workspace/ib-any-integration/CMSSW_8_1_X_2016-09-17-1100/src/EventFilter/CSCRawToDigi/interface/CSCEventData.h:33,
                 from /build/cmsbuild/jenkins-workarea/workspace/ib-any-integration/CMSSW_8_1_X_2016-09-17-1100/src/EventFilter/CSCRawToDigi/interface/CSCDDUEventData.h:11,
                 from /build/cmsbuild/jenkins-workarea/workspace/ib-any-integration/CMSSW_8_1_X_2016-09-17-1100/src/EventFilter/CSCRawToDigi/src/CSCDDUDataItr.cc:2:
/build/cmsbuild/jenkins-workarea/workspace/ib-any-integration/CMSSW_8_1_X_2016-09-17-1100/src/EventFilter/CSCRawToDigi/interface/CSCCLCTData.h:38:15: error: 'CSCComparatorDigi' was not declared in this scope
   std::vector<CSCComparatorDigi> comparatorDigis(int layer);
               ^
/build/cmsbuild/jenkins-workarea/workspace/ib-any-integration/CMSSW_8_1_X_2016-09-17-1100/src/EventFilter/CSCRawToDigi/interface/CSCCLCTData.h:38:32: error: template argument 1 is invalid
   std::vector<CSCComparatorDigi> comparatorDigis(int layer);
                                ^
</pre>