@cmsbuild 2016-09-18
The tests are being triggered in jenkins.