@slava77 2016-09-20
Here are some technical performance numbers for step3 in ttbar PU35, using 70 events, configured without DQM and Validation to get some expectations for 2017 production needs (the baseline is CMSSW_8_1_X_2016-09-19-1100):
- baseline 25202 (2016) vs 10224 (2017)
    - CPU increases by about 15% (16.3 to 18.7 s/evt), essentially all due to tracking (both in the iterative and in electron/muon-related, as well as downstream track-dependent algorithms); hbhe local reco is ~20% faster here.
    - memory is roughly unchanged peak around 1.8 GiB in single-thread running; small increase from 2.6 to 2.7 GiB in 4-thread run. The test is likely too short to properly probe the memory peak.
    - size on disk increases by about 9% in RECO (combination of tracks, siPixelClusters), about 11% in AOD (combination of generalTracks, pfCands and jets); MINIAOD reduced by 5% (~+5% from packedCands and jets is balanced out by 10% decrease in patTrigger info, which should go back up once real trigger setup is available)
- baseline 10224 (2017 setup with GEN-SIM inputs from relval) vs this PR 10224 (GEN-SIM is generated as a part of the tests)
    - CPU increases by about 27% (18.7 to 23.8 s/evt), essentially all in hbheprereco (from 0.65 s to 6.4 s/evt) and hbhereco (from 0.04 to 0.3 s/evt)
    - memory increases by 0.3 GiB in signle-thread running;  decrease from 2.7 to 2.5 GiB in 4-thread running. The test is likely too short to properly probe the memory peak.
    - size on disk in RECO is up by about 18% (dominated by 200kB in HBHEChannelInfosSorted_hbheprereco  and +90 kB in HBHERecHitsSorted_hbheprereco)
        -  **consider to remove HBHEChannelInfos** else you will limit RECO availability to all 
    - size on disk in AOD is down by 5%. A half of this decrease is driven by factor of ~10 reduction in HF and HBHE reducedHcalRecHits (if I'm not mistaken, this is temporary and a large fraction may come back); the other half is from some (~3%) reduction in pfCands and jets and tracks (unclear if tracks change is a fluctuation or an effect of geometry etc changes).

More on the changes in baseline vs this PR for 10224:
- ZDC hits are empty: this is probably OK for now, but it should be fixed for the post-2017 HI runs (or possible other special runs)
    -  (note that Castor is still in the simulation)
- hbhe reco hit count goes up by ~5 (to about 7.5E3 hits per event), which looks more than just the total channel count increase. Are the zero suppression thresholds appropriately in place?
- horeco hit count is down by about 8: this this the result of some fix in SIM (recall that my baseline is GEN-SIM from relval), or is it from the "Dev" conditions?
- HcalHaloData phiWedgeCollection size is down by a factor of 3: unclear if it's a reduction in noise or if the algorithm will need retuning
- hcal noise RBX type1 (e.g. hRBXEnergyType1) DQM plots are empty. I guess it's related to the noise flags possibly not all set correctly (there is a PR with fixes).

I'll post notes on changes with less than ~order of magnitude change later.